requirements.txt
# 基础依赖（必需）
fastapi>=0.111.1
uvicorn[standard]
python-multipart
websockets
python-dotenv

# 数据处理
numpy<=2.3.3

# 机器学习框架
# CPU版本（开发/测试）
torch>=2.0.0; platform_machine!="x86_64" or sys_platform=="darwin"
torchaudio>=2.0.0; platform_machine!="x86_64" or sys_platform=="darwin"

# GPU版本（生产环境），通过环境变量控制
# 使用时请根据CUDA版本选择合适的index-url
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# 音频处理
pydub>=0.25.1

# SenseVoice相关
modelscope
huggingface_hub
funasr>=1.1.3

# 可选依赖
gradio
